chore: restructure repo (#56)
