+++ include/common/firmware/motor.h
