+++ pipettes/firmware/CMakeLists.txt
@@ -8,7 +8,6 @@ add_STM32G4_freertos("Pipettes")
 set(PIPETTE_FW_LINTABLE_SRCS
         ${CMAKE_CURRENT_SOURCE_DIR}/main.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/freertos_idle_timer_task.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/freertos_comms_task.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/uart_comms.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/uart_task.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/can_task.cpp
