+++ pipettes/tests/test_message_handler.cpp
@@ -1,5 +1,5 @@
 #include "catch2/catch.hpp"
-#include "common/hal/io.hpp"
+#include "common/core/io.hpp"
 #include "pipettes/core/pipette_messages.hpp"
 #include "pipettes/core/uart_message_handler.hpp"
 #include "pipettes/tests/test_motor_control.hpp"
